+++ CHANGELOG.md
@@ -5,6 +5,12 @@ However, as this is still an experimental library, breaking changes may occur wi
 
 The official release will have a defined and more stable API. If you are already relying on a particular API, please let me know.
 
+## 0.4.3
+
+* Improvements:
+  * Now providing events `:step-result-updated` and `:step-finished` on event-bus for use by other components like
+    runners and persistence components
+
 ## 0.4.2
 
 * Improvements: